refactor: 컴포넌트 이름 변경 & 캐릭터 아이템 배경색 추가
