+++ src/components/scenario/CharacterHeader/CharacterHeader.style.ts
@@ -4,7 +4,7 @@ export const CharacterHeaderLeftSide = styled.div`
   display: flex;
   gap: 8px;
 `;
-export const CharacterHeader = styled.div<{ isChecked: boolean; isEdit: boolean }>`
+export const CharacterHeader = styled.div<{ isChecked: boolean; isEdit: boolean; idx?: number }>`
   display: flex;
   padding: 30px 20px;
   align-items: center;
@@ -13,7 +13,9 @@ export const CharacterHeader = styled.div<{ isChecked: boolean; isEdit: boolean 
   justify-content: space-between;
   border-radius: 4px;
   border-top: ${({ isEdit }) => (isEdit ? '1px solid rgba(217, 230, 255, 0.2)' : 'none')};
-  background: ${({ isChecked }) => (isChecked ? 'rgba(217, 230, 255, 0.05)' : 'inherit')};
+
+  background: ${({ idx, isChecked }) =>
+    isChecked ? 'rgba(217, 230, 255, 0.05)' : idx === 0 ? 'linear-gradient(76deg, #001945 0%, #000714 100.13%)' : 'inherit'};
 `;
 
 export const CharacterHeaderRightSide = styled.div`
