+++ src/components/scenario/CharacterList/CharacterList.style.ts
