+++ src/components/scenario/CheckBox/CheckBox.tsx
@@ -14,27 +14,27 @@ interface ICheckBoxProps {
 export default function CheckBox({ characterId, scenarioId, isAllCheckBox = false }: ICheckBoxProps) {
   const dispatch = useDispatch<TAppDispatch>();
 
-  //  체크 여부 판단
+  // 체크 여부 판단
   const isChecked: boolean = useSelector((state: TRootState) => {
     if (isAllCheckBox) {
-      //ALL
+      // ALL
       return state.scenario.characters.every((char) => char.isChecked && char.scenarios.every((scn) => scn.isChecked));
     }
     if (scenarioId) {
-      //시나리오
+      // 시나리오
       const character = state.scenario.characters.find((char) => char.id === characterId);
       const scenario = character?.scenarios.find((scn) => scn.id === scenarioId);
       return scenario?.isChecked ?? false;
     }
     if (characterId) {
-      //역할
+      // 역할
       const character = state.scenario.characters.find((char) => char.id === characterId);
       return character?.isChecked ?? false;
     }
     return false;
   });
 
-  //  클릭 함수
+  // 클릭 함수
   const handleCheckBoxClick = (): void => {
     if (isAllCheckBox) {
       dispatch(toggleAll());
